created search router, getProductsByCategory is working, but getProductsBySearch still needs debugging
