+++ package.json
@@ -17,8 +17,9 @@
     "express-favicon": "^2.0.4",
     "http-status-codes": "^2.2.0",
     "jsonwebtoken": "^9.0.0",
-    "mongodb": "^5.5.0",
-    "mongoose": "^7.0.1",
+    "mongodb": "^5.6.0",
+    "mongoose": "^7.2.3",
+    "mongosh": "^1.10.0",
     "morgan": "^1.10.0",
     "nodemailer": "^6.9.3"
   },
