+++ src/app.js
@@ -9,7 +9,8 @@ const authenticateUser = require('./middleware/authentication')
 //routers
 const mainRouter = require('./routes/mainRouter.js');
 const authRouter = require('./routes/auth');
-const productRouter = require('./routes/products')
+const productRouter = require('./routes/products');
+const searchRouter= require('./routes/search');
 
 
 //error handler
@@ -28,6 +29,7 @@ app.use(favicon(__dirname + '/public/favicon.ico'));
 app.use('/api/v1', mainRouter);
 app.use('/api/v1/auth', authRouter); 
 app.use('/api/v1/products', authenticateUser, productRouter)
+app.use('/api/v1/search', authenticateUser, searchRouter)
 
 app.use(notFoundMiddleware);
 app.use(errorHandlerMiddleware);
