+++ src/routes/products.js
@@ -8,16 +8,10 @@ const {
 	getAllProducts,
 	updateProduct,
 	getProduct,
-	getProductsBySearch,
-	getProductsByFilter,
 } = require('../controllers/products')
 
 router.route('/').post(createProduct).get(getAllProducts)
 
 router.route('/:id').get(getProduct).delete(deleteProduct).patch(updateProduct)
-
-router.route('/search').get(getProductsBySearch)
-
-router.route('/filter').get(getProductsByFilter)
 
 module.exports = router
